# docs/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cahn Family Task API",
    "description": "API voor de Cahn Family Task Assistant - eerlijke takenverdeling voor het gezin. Taken: uitruimen ochtend (1x/week), uitruimen avond (2x/week), inruimen (2x/week), dekken (2x/week), karton/papier (1x/week), glas (1x/3 weken), koken (1x/3 weken).",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://cahn-family-assistent.vercel.app",
      "description": "Production (Vercel)"
    }
  ],
  "paths": {
    "/api/tasks": {
      "get": {
        "operationId": "listTasks",
        "summary": "Haal alle taken op met hun configuratie",
        "description": "Geeft een lijst van alle taken met hun volledige beschrijving en frequentie (per kind per week)",
        "responses": {
          "200": {
            "description": "Lijst van alle taken",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string", "description": "Interne naam van de taak" },
                      "display_name": { "type": "string", "description": "Vriendelijke naam" },
                      "description": { "type": "string", "description": "Volledige beschrijving wat de taak inhoudt" },
                      "weekly_target": { "type": "integer", "description": "Totaal aantal keer per week (alle kinderen samen)" },
                      "per_child_target": { "type": "integer", "description": "Aantal keer per kind per week" },
                      "rotation_weeks": { "type": "integer", "description": "Elke hoeveel weken (1=wekelijks, 3=om de 3 weken)" },
                      "time_of_day": { "type": "string", "description": "Tijdstip: ochtend, middag, avond" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/suggest/{task_name}": {
      "get": {
        "operationId": "suggestForTask",
        "summary": "Vraag wie aan de beurt is voor een taak",
        "description": "Berekent op basis van eerlijke verdeling wie de taak zou moeten doen",
        "parameters": [
          {
            "name": "task_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Naam van de taak: uitruimen_ochtend, uitruimen_avond, inruimen, dekken, karton_papier, glas, koken"
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestie wie de taak moet doen",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggested": { "type": "string", "description": "Naam van gesuggereerd gezinslid" },
                    "reason": { "type": "string", "description": "Uitleg waarom deze persoon aan de beurt is" },
                    "scores": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "total_this_week": { "type": "integer" },
                          "score": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/complete": {
      "post": {
        "operationId": "completeTask",
        "summary": "Registreer dat iemand een taak heeft gedaan",
        "description": "Voor één taak. Bij MEERDERE taken: gebruik /api/complete/bulk om verwarring te voorkomen.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "member_name": {
                    "type": "string",
                    "description": "Naam: Nora, Linde of Fenna"
                  },
                  "task_name": {
                    "type": "string",
                    "description": "Taak: dekken, inruimen, uitruimen, uitruimen_ochtend, uitruimen_avond, karton_papier, glas, koken"
                  }
                },
                "required": ["member_name", "task_name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Taak succesvol geregistreerd",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "completion_id": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/complete/bulk": {
      "post": {
        "operationId": "completeTasksBulk",
        "summary": "Registreer meerdere taken in één keer",
        "description": "GEBRUIK DIT bij meerdere taken tegelijk. Voorkomt verwarring over wie wat deed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completions": {
                    "type": "array",
                    "description": "Lijst van taken met expliciet wie wat deed",
                    "items": {
                      "type": "object",
                      "properties": {
                        "member_name": { "type": "string", "description": "Naam: Nora, Linde of Fenna" },
                        "task_name": { "type": "string", "description": "Taak die deze persoon deed" }
                      },
                      "required": ["member_name", "task_name"]
                    }
                  }
                },
                "required": ["completions"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Taken geregistreerd",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "results": { "type": "array" },
                    "errors": { "type": "array" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/summary": {
      "get": {
        "operationId": "getWeeklySummary",
        "summary": "Haal het weekoverzicht op met taken per persoon",
        "responses": {
          "200": {
            "description": "Overzicht van taken deze week",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Nora": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer", "description": "Totaal aantal taken" },
                        "tasks": { "type": "object", "description": "Taken per type" }
                      }
                    },
                    "Linde": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer", "description": "Totaal aantal taken" },
                        "tasks": { "type": "object", "description": "Taken per type" }
                      }
                    },
                    "Fenna": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer", "description": "Totaal aantal taken" },
                        "tasks": { "type": "object", "description": "Taken per type" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schedule": {
      "get": {
        "operationId": "getWeekSchedule",
        "summary": "Haal het weekrooster op met per-dag planning",
        "description": "Geeft een SUGGESTIE voor wie welke taak zou kunnen doen per dag. Dit is GEEN verplichting - kinderen mogen onderling wisselen of andere taken doen. Registreer altijd wat er DAADWERKELIJK is gedaan, niet wat gepland stond. Bevat ook een ASCII/emoji overzicht dat direct getoond kan worden.",
        "responses": {
          "200": {
            "description": "Weekrooster met ASCII overzicht",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "week_number": { "type": "integer", "description": "ISO week nummer" },
                    "week_start": { "type": "string", "description": "Startdatum van de week (maandag)" },
                    "schedule": {
                      "type": "object",
                      "description": "Per dag de taken met toegewezen persoon en status"
                    },
                    "ascii_overview": { "type": "string", "description": "ASCII/emoji weekoverzicht om direct te tonen" },
                    "member_totals": {
                      "type": "object",
                      "description": "Totaal aantal taken per persoon deze week",
                      "properties": {
                        "Nora": { "type": "integer" },
                        "Linde": { "type": "integer" },
                        "Fenna": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/undo": {
      "post": {
        "operationId": "undoLastTask",
        "summary": "Maak de laatst voltooide taak ongedaan",
        "description": "Verwijdert de laatst geregistreerde taak voor een gezinslid. Handig als iemand per ongeluk de verkeerde taak heeft afgevinkt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "member_name": {
                    "type": "string",
                    "description": "Naam van het gezinslid: Nora, Linde of Fenna"
                  }
                },
                "required": ["member_name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultaat van de undo actie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "undone_task": { "type": "string", "description": "Naam van de ongedaan gemaakte taak (alleen bij success)" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/absence": {
      "post": {
        "operationId": "registerAbsence",
        "summary": "Registreer dat iemand afwezig is",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "member_name": { "type": "string" },
                  "start_date": { "type": "string", "format": "date" },
                  "end_date": { "type": "string", "format": "date" },
                  "reason": { "type": "string" }
                },
                "required": ["member_name", "start_date", "end_date"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Afwezigheid geregistreerd"
          }
        }
      }
    },
    "/api/swap/request": {
      "post": {
        "operationId": "requestSwap",
        "summary": "Vraag een taakruil aan met een ander gezinslid",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requester_name": { "type": "string" },
                  "target_name": { "type": "string" },
                  "task_name": { "type": "string" },
                  "swap_date": { "type": "string", "format": "date" }
                },
                "required": ["requester_name", "target_name", "task_name", "swap_date"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ruil verzoek aangemaakt"
          }
        }
      }
    },
    "/api/swap/respond": {
      "post": {
        "operationId": "respondToSwap",
        "summary": "Accepteer of weiger een ruil verzoek",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "swap_id": { "type": "string" },
                  "accept": { "type": "boolean" }
                },
                "required": ["swap_id", "accept"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ruil verzoek beantwoord"
          }
        }
      }
    },
    "/api/swaps/pending/{member_name}": {
      "get": {
        "operationId": "getPendingSwaps",
        "summary": "Haal openstaande ruil verzoeken op voor een gezinslid",
        "parameters": [
          {
            "name": "member_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lijst met openstaande ruil verzoeken"
          }
        }
      }
    }
  }
}